requirements.txt
openai
python-dotenv
praw
panel
